+++ chrysom/client_test.go
@@ -49,6 +49,7 @@ func TestValidateConfig(t *testing.T) {
 		PullInterval:    time.Second * 5,
 		Logger:          log.NewNopLogger(),
 		Address:         "http://awesome-argus-hostname.io",
+		Bucket:          "bucket-name",
 		MetricsProvider: provider.NewDiscardProvider(),
 	}
 
@@ -58,40 +59,40 @@ func TestValidateConfig(t *testing.T) {
 		PullInterval:    time.Hour * 24,
 		Address:         "http://legit-argus-hostname.io",
 		Auth:            Auth{},
-		MetricsProvider: provider.NewDiscardProvider(),
+		MetricsProvider: provider.NewExpvarProvider(),
+		Bucket:          "amazing-bucket",
 		Logger:          log.NewJSONLogger(ioutil.Discard),
 	}
 
 	tcs := []testCase{
 		{
-			Description: "All default values",
-			Input: &ClientConfig{
-				Address:         "http://awesome-argus-hostname.io",
-				MetricsProvider: provider.NewDiscardProvider(),
-			},
-			ExpectedConfig: allDefaultsCaseConfig,
-		},
-
-		{
-			Description: "No metrics provider",
-			Input: &ClientConfig{
-				Address: "http://awesome-argus-hostname.io",
-			},
-			ExpectedErr: ErrUndefinedMetricsProvider,
-		},
-		{
 			Description: "No address",
 			Input: &ClientConfig{
-				MetricsProvider: provider.NewDiscardProvider(),
+				Bucket: "bucket-name",
 			},
 			ExpectedErr: ErrAddressEmpty,
 		},
-
+		{
+			Description: "No bucket",
+			Input: &ClientConfig{
+				Address: "http://awesome-argus-hostname.io",
+			},
+			ExpectedErr: ErrBucketEmpty,
+		},
+		{
+			Description: "All default values",
+			Input: &ClientConfig{
+				Address: "http://awesome-argus-hostname.io",
+				Bucket:  "bucket-name",
+			},
+			ExpectedConfig: allDefaultsCaseConfig,
+		},
 		{
 			Description: "All defined",
 			Input: &ClientConfig{
-				MetricsProvider: provider.NewDiscardProvider(),
+				MetricsProvider: provider.NewExpvarProvider(),
 				Address:         "http://legit-argus-hostname.io",
+				Bucket:          "amazing-bucket",
 				HTTPClient:      myAmazingClient,
 				PullInterval:    time.Hour * 24,
 				Logger:          log.NewJSONLogger(ioutil.Discard),
@@ -176,6 +177,7 @@ func TestSendRequest(t *testing.T) {
 			client, err := NewClient(ClientConfig{
 				HTTPClient:      server.Client(),
 				Address:         "http://argus-hostname.io",
+				Bucket:          "bucket-name",
 				MetricsProvider: provider.NewDiscardProvider(),
 			})
 
@@ -206,7 +208,6 @@ func TestGetItems(t *testing.T) {
 		Description           string
 		ResponsePayload       []byte
 		ResponseCode          int
-		ShouldEraseBucket     bool
 		ShouldMakeRequestFail bool
 		ShouldDoRequestFail   bool
 		ExpectedErr           error
@@ -214,11 +215,6 @@ func TestGetItems(t *testing.T) {
 	}
 
 	tcs := []testCase{
-		{
-			Description:       "Bucket is required",
-			ShouldEraseBucket: true,
-			ExpectedErr:       ErrBucketEmpty,
-		},
 		{
 
 			Description:           "Make request fails",
@@ -268,10 +264,6 @@ func TestGetItems(t *testing.T) {
 				owner   = "owner-name"
 			)
 
-			if tc.ShouldEraseBucket {
-				bucket = ""
-			}
-
 			server := httptest.NewServer(http.HandlerFunc(func(rw http.ResponseWriter, r *http.Request) {
 				assert.Equal(http.MethodGet, r.Method)
 				assert.Equal(owner, r.Header.Get(store.ItemOwnerHeaderKey))
@@ -284,6 +276,7 @@ func TestGetItems(t *testing.T) {
 			client, err := NewClient(ClientConfig{
 				HTTPClient:      server.Client(),
 				Address:         server.URL,
+				Bucket:          bucket,
 				MetricsProvider: provider.NewDiscardProvider(),
 			})
 
@@ -297,7 +290,7 @@ func TestGetItems(t *testing.T) {
 				client.storeBaseURL = failingURL
 			}
 
-			output, err := client.GetItems(bucket, "owner-name")
+			output, err := client.GetItems(owner)
 
 			assert.True(errors.Is(err, tc.ExpectedErr))
 			if tc.ExpectedErr == nil {
@@ -313,7 +306,6 @@ func TestPushItem(t *testing.T) {
 		Item                  model.Item
 		Owner                 string
 		ResponseCode          int
-		ShouldEraseID         bool
 		ShouldEraseBucket     bool
 		ShouldRespNonSuccess  bool
 		ShouldMakeRequestFail bool
@@ -333,26 +325,9 @@ func TestPushItem(t *testing.T) {
 
 	tcs := []testCase{
 		{
-			Description:       "Bucket is required",
-			Item:              validItem,
-			ShouldEraseBucket: true,
-			ExpectedErr:       ErrBucketEmpty,
-		},
-		{
-			Description:   "Item ID Missing",
-			Item:          validItem,
-			ShouldEraseID: true,
-			ExpectedErr:   ErrItemIDEmpty,
-		},
-		{
-			Description: "Item ID Missing from payload",
-			Item:        model.Item{Data: validItem.Data},
+			Description: "Item ID Missing",
+			Item:        model.Item{Data: map[string]interface{}{}},
 			ExpectedErr: ErrItemIDEmpty,
-		},
-		{
-			Description: "Item ID Mismatch",
-			Item:        model.Item{ID: "752f10c83610ebca1a059c0bae8255eba2f95be4d1d7bcfa89d7248a82d9f119", Data: validItem.Data},
-			ExpectedErr: ErrItemIDMismatch,
 		},
 		{
 			Description: "Item Data missing",
@@ -438,6 +413,7 @@ func TestPushItem(t *testing.T) {
 			client, err := NewClient(ClientConfig{
 				HTTPClient:      server.Client(),
 				Address:         server.URL,
+				Bucket:          bucket,
 				MetricsProvider: provider.NewDiscardProvider(),
 			})
 
@@ -453,12 +429,8 @@ func TestPushItem(t *testing.T) {
 				bucket = ""
 			}
 
-			if tc.ShouldEraseID {
-				id = ""
-			}
-
 			require.Nil(err)
-			output, err := client.PushItem(id, bucket, tc.Owner, tc.Item)
+			output, err := client.PushItem(tc.Owner, tc.Item)
 
 			if tc.ExpectedErr == nil {
 				assert.EqualValues(tc.ExpectedOutput, output)
@@ -475,8 +447,6 @@ func TestRemoveItem(t *testing.T) {
 		ResponsePayload       []byte
 		ResponseCode          int
 		Owner                 string
-		ShouldEraseBucket     bool
-		ShouldEraseID         bool
 		ShouldRespNonSuccess  bool
 		ShouldMakeRequestFail bool
 		ShouldDoRequestFail   bool
@@ -485,16 +455,6 @@ func TestRemoveItem(t *testing.T) {
 	}
 
 	tcs := []testCase{
-		{
-			Description:       "Bucket is required",
-			ShouldEraseBucket: true,
-			ExpectedErr:       ErrBucketEmpty,
-		},
-		{
-			Description:   "Item ID Missing",
-			ShouldEraseID: true,
-			ExpectedErr:   ErrItemIDEmpty,
-		},
 		{
 			Description:           "Make request fails",
 			ShouldMakeRequestFail: true,
@@ -552,6 +512,7 @@ func TestRemoveItem(t *testing.T) {
 			client, err := NewClient(ClientConfig{
 				HTTPClient:      server.Client(),
 				Address:         server.URL,
+				Bucket:          bucket,
 				MetricsProvider: provider.NewDiscardProvider(),
 			})
 
@@ -563,16 +524,8 @@ func TestRemoveItem(t *testing.T) {
 				client.storeBaseURL = failingURL
 			}
 
-			if tc.ShouldEraseID {
-				id = ""
-			}
-
-			if tc.ShouldEraseBucket {
-				bucket = ""
-			}
-
 			require.Nil(err)
-			output, err := client.RemoveItem(id, bucket, tc.Owner)
+			output, err := client.RemoveItem(id, tc.Owner)
 
 			if tc.ExpectedErr == nil {
 				assert.EqualValues(tc.ExpectedOutput, output)