+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Changed
+- Better validation and documentation for Argus config. [#141](https://github.com/xmidt-org/argus/issues/141)
+- Make bucket a required field for the client not just the listener. [#141](https://github.com/xmidt-org/argus/issues/141)
+
 ### Fixed
 - Super user getAll requests should be filtered by owner when one is provided. [#136](https://github.com/xmidt-org/argus/pull/136)