+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Changed
-- Better validation and documentation for Argus config. [#141](https://github.com/xmidt-org/argus/issues/141)
-- Make bucket a required field for the client not just the listener. [#141](https://github.com/xmidt-org/argus/issues/141)
+- Better validation and documentation for Argus config. [#141](https://github.com/xmidt-org/argus/pull/141)
+- Make bucket a required field for the client not just the listener. [#141](https://github.com/xmidt-org/argus/pull/141)
+- Group listener-only config for client. [#142] (https://github.com/xmidt-org/argus/pull/142)
 
 ### Fixed
 - Super user getAll requests should be filtered by owner when one is provided. [#136](https://github.com/xmidt-org/argus/pull/136)